requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.3
dj-database-url==1.2.0
dj-static==0.0.6
Django==4.1.3
django-rest-knox==4.2.0
djangorestframework==3.14.0
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.1.0
pip==22.3.1
psycopg2==2.9.5
pycparser==2.21
pytz==2022.6
setuptools==49.2.1
sqlparse==0.4.3
static3==0.7.0
tzdata==2022.6
whitenoise==6.3.0
django-heroku==0.3.1